requirements.txt
aioboto3==12
fsspec
openai~=1.14
pandas==2
pydantic==2
s3fs
streamlit>=1.36
tiktoken==0.7.0
langchain-community~=0.2.15
langchain~=0.2.15
langchain_core~=0.2.15
langchain-openai~=0.1.23
cfn_flip
cfn-lint
ipykernel
notebook
openpyxl
beautifulsoup4
pyarrow
litellm
chromadb>=0.5.5
pypdf>=4.2.0
python-dotenv>=1.0.1
langchain-text-splitters~=0.2.2
sentence-transformers>=2.5.1
